refactor: rename Rating to Rating.Stars
BREAKING CHANGE: Rating is not a component anymore, now it is an object
with both RatingStars and RatingThumbs as props (Rating.Stars and
Rating.Thumbs) respectively
