+++ packages/picasso/src/RatingStars/__image_snapshots__/rating-custom-icon-snap.png
