+++ packages/picasso/src/RatingStars/__image_snapshots__/rating-default-snap.png
