+++ packages/picasso/src/RatingStars/__image_snapshots__/rating-non-interactive-snap.png
