+++ packages/picasso/src/RatingStars/__image_snapshots__/rating-readonly-snap.png
