+++ packages/picasso/src/RatingStars/index.ts
@@ -1,0 +1,6 @@
+import { OmitInternalProps } from '@toptal/picasso-shared'
+
+import { Props } from './RatingStars'
+
+export { default } from './RatingStars'
+export type RatingStarsProps = OmitInternalProps<Props>
