+++ packages/picasso/src/RatingStars/story/index.jsx
@@ -1,5 +1,5 @@
 import ratingThumbsStory from '../../RatingThumbs/story'
-import Rating from '../Rating'
+import RatingStars from '../RatingStars'
 import PicassoBook from '~/.storybook/components/PicassoBook'
 
 const page = PicassoBook.section('Components').createPage(
@@ -16,26 +16,26 @@ const page = PicassoBook.section('Components').createPage(
 
 page
   .createTabChapter('Props')
-  .addComponentDocs({ component: Rating, name: 'Rating' })
+  .addComponentDocs({ component: RatingStars, name: 'Rating.Stars' })
   .addComponentDocs(ratingThumbsStory.componentDocs)
 
 page
   .createChapter()
-  .addExample('Rating/story/Default.example.tsx', {
+  .addExample('RatingStars/story/Default.example.tsx', {
     title: 'Default',
     description: 'Default behavior'
   }) // picasso-skip-visuals
-  .addExample('Rating/story/NonInteractive.example.tsx', {
+  .addExample('RatingStars/story/NonInteractive.example.tsx', {
     title: 'Non Interactive',
     description:
       'The rating can be used in non-interactive mode. This mode prohibits changing its value.'
   }) // picasso-skip-visuals
-  .addExample('Rating/story/Sizes.example.tsx', {
+  .addExample('RatingStars/story/Sizes.example.tsx', {
     title: 'Sizes example',
     description:
       'The rating component has two sizes - small (default) and large. It can be set via size parameter.'
   }) // picasso-skip-visuals
-  .addExample('Rating/story/CustomIcons.example.tsx', {
+  .addExample('RatingStars/story/CustomIcons.example.tsx', {
     title: 'Custom Icon',
     description:
       'The icons are fully customizable. You can provide a custom icon to render or add a tooltip to the default one.'
