+++ packages/picasso/src/RatingStars/styles.ts
